+++ internal/app/http/common.go
@@ -275,10 +275,10 @@ func convertParams(params any) *model.Filters {
 		&model.FilterOptions{NameFilterType: model.PrefixMatch})
 }
 
-func toInventoryResponse(ctx context.Context, res model.SearchResult) server.InventoryResponse {
+func toInventoryResponse(ctx context.Context, res model.SearchResult, page, pageSize int) server.InventoryResponse {
 	mapper := NewMapper(ctx)
 
-	meta := mapper.GetInventoryMeta(res)
+	meta := mapper.GetInventoryMeta(res, page, pageSize)
 	inv := mapper.GetInventoryData(res.Entries)
 	resp := server.InventoryResponse{
 		Meta: &meta,