+++ internal/app/http/server/models.gen.go
@@ -155,7 +155,14 @@ type Meta struct {
 
 // MetaPage defines model for MetaPage.
 type MetaPage struct {
-	Elements int `json:"elements"`
+	// PageNumber current page number
+	PageNumber *int `json:"pageNumber,omitempty"`
+
+	// PageSize size of a page
+	PageSize *int `json:"pageSize,omitempty"`
+
+	// TotalElements total number of elements in the catalog
+	TotalElements *int `json:"totalElements,omitempty"`
 }
 
 // ModelVersion defines model for ModelVersion.
@@ -288,6 +295,22 @@ type GetInventoryParams struct {
 	// FilterLatest Filters the inventory to return only the latest versions of data.
 	// If this filter is present in the URL (e.g., `?filter.latest`), only the latest versions will be returned. If it's not present, all versions will be included.
 	FilterLatest *bool `form:"filter.latest,omitempty" json:"filter.latest,omitempty"`
+
+	// Page Page number for pagination (starting from 1)
+	//
+	// - If `pageSize` is provided along with `page`, both values are used for pagination.
+	// - If `pageSize` is provided but `page` is *not* provided, `page` will default to `1`.
+	// - If `pageSize` is *not* provided but `page` *is* provided, `pageSize` will default to `100`.
+	// - If *neither* `page` nor `pageSize` are provided, no pagination will be applied, and both parameters will effectively be treated as `0` (returning all results).
+	Page *int `form:"page,omitempty" json:"page,omitempty"`
+
+	// PageSize Number of InventoryEntries per page for pagination.
+	//
+	// - If `pageSize` is provided along with `page`, both values are used for pagination.
+	// - If `pageSize` is provided but `page` is *not* provided, `page` will default to `1`.
+	// - If `pageSize` is *not* provided but `page` *is* provided, `pageSize` will default to `100`.
+	// - If *neither* `page` nor `pageSize` are provided, no pagination will be applied, and both parameters will effectively be treated as `0` (returning all results).
+	PageSize *int `form:"pageSize,omitempty" json:"pageSize,omitempty"`
 
 	// Search Searches the inventory for TMs that match the search query. Accepts queries in bleve search engine syntax.
 	// Is mutually exclusive with filters.